impl readOnlyStakingPool
